Fix license info for the binaries
Analysis of Jaroslav Tulach shows, that the whole nb-javac source
is covered by the Classpath Exception:

https://lists.apache.org/thread.html/r821d9e9fdc8d9fd5663e7c326d25e4626e1a27eb13e45f4d639ea199%40%3Cdev.netbeans.apache.org%3E

I did an independend check and came to the same conclusion. The problem
is, that the license file says, that it might only apply to a subset
of the source files and thus leaves users of the binary in a bad
situation, as they have to speculate, whether or not the binary was
build only from source files covered by the Class Path exception.

This change adds a license file only for the binary build artifacts,
which clarifies that the build binary as a whole is covered by the
Classpath Exception. Only the two binary jars are affected. The
license for the doc and source artifact are unchanged.


Signed-of-by: Matthias Bläsing <mblaesing@doppel-helix.eu>